requirements.txt
Django==2.2.1
django-bootstrap-breadcrumbs==0.9.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
django-view-breadcrumbs==0.0.9
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
Pillow==6.0.0
django-bootstrap-select 0.1.3
pytz==2019.1
reportlab==3.5.21
six==1.12.0
sqlparse==0.3.0
manifest-tool 1.4.8
django-mathfilters 0.4.0
